Added facility rollup module
