+++ db/migrate/20201102162746_add_facility_view_in_privileges.rb
@@ -1,0 +1,5 @@
+class AddFacilityViewInPrivileges < ActiveRecord::Migration[5.2]
+  def change
+    add_column :privileges, :facility_manager_view, :string, default: "R"
+  end
+end
